refactor: extract useHasHydrated to shared hooks
